requirements.txt
# for windows only
transformers
python-dotenv
accelerate
scipy
pyaudio
numpy
CodingDrone
keyboard
